requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-barcode==0.15.1
Pillow==10.0.1
Werkzeug==2.3.7
gunicorn==21.2.0
bleach==6.0.0